// src/app/models/transmittalResponse.ts
/**
 * Revenue Management System
 * Florida Fish and Wildlife Conservation Commision - Revenue Management System.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: apiteam@fwc.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface TransmittalResponse { 
    transmittalNumber: number;
    transmittalDate: string;
    transmittalTotalCount?: number;
    transmittalTotal?: number;
    transmittalStatus: string;
}