// src/app/models/createTransmittalDetailRequest.ts
/**
 * Revenue Management System
 * Florida Fish and Wildlife Conservation Commision - Revenue Management System.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: apiteam@fwc.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CreateTransmittalDetailRequest { 
    checkNumber?: string;
    firstName?: string;
    lastName?: string;
    companyName?: string;
    checkAmount?: number;
    cashListing?: string;
    comments?: string;
}